Started reorganizing code
Preliminary attempt to split code into organized files. I thought I understood how definitions, headers and external variables worked, but there are multiple issues I need to resolve.
