+++ bin/Debug/New folder/libstdc++-6.dll
