+++ clicks.h
@@ -1,0 +1,7 @@
+#ifndef CLICKS_H_INCLUDED
+#define CLICKS_H_INCLUDED
+
+void writeClick(char *button); // registers which click and when into an array
+void checkClick(char *button); // checks if right or left button was clicked
+
+#endif // CLICKS_H_INCLUDED
