+++ menu.h
@@ -1,0 +1,8 @@
+#ifndef MENU_H_INCLUDED
+#define MENU_H_INCLUDED
+
+void setUnits(); // assigns units
+void setTime(); // assigns time
+void checkLoop(); // checks if you want to restart
+
+#endif // MENU_H_INCLUDED
